Rename logging modules and update imports
