+++ src/app/api/main.py
@@ -23,7 +23,7 @@ from app.api.routes.metrics import router as metrics_router
 from app.api.routes.review import router as review_router
 from app.api.routes.status import router as status_router
 from app.core.config import settings
-from app.core.loging import get_logger
+from app.core.logging import get_logger
 from app.observability.app_insights import app_insights
 from app.observability.prometheus import instrument_app
