+++ src/app/memory/storage.py
@@ -15,7 +15,7 @@ from typing import Any, Literal
 import aiosqlite
 
 from app.core.config import MAX_MEMORY, MAX_TOTAL_MEMORY
-from src.app.core.loging import get_logger
+from src.app.core.logging import get_logger
 
 logger = get_logger(__name__)
